Added update and destroy function for dogs. Made dog controller
